feat(media): create separate pages for media and settings
